+++ src/constants/index.ts
@@ -1,8 +1,14 @@
 export const DEFAULT_ALPHABET =
   "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
 
+export const BASE32_CROCKFORD = "0123456789ABCDEFGHJKMNPQRSTVWXYZ";
+
 export const DEFAULT_SIZE = 24;
 
 export const DEFAULT_SEPARATOR = "_";
 
 export const MAX_SIZE = 4096;
+
+export const DEFAULT_SORTABLE_RANDOM_SIZE = 16;
+
+export const SORTABLE_TIME_MAX = 2 ** 48 - 1;