+++ src/index.ts
@@ -1,8 +1,18 @@
-export { createId, id } from "./generate.js";
-export { ensureUnique } from "./ensure-unique.js";
-export { template } from "./template.js";
-export { getPrefix, isId } from "./validate.js";
+export { BASE32_CROCKFORD } from "./constants/index.js";
+export { createId, id } from "./generators/generate.js";
+export {
+  createSortableId,
+  getTimestamp,
+  sortableId,
+} from "./generators/sortable.js";
+export { template } from "./generators/template.js";
+export { ensureUnique } from "./utils/ensure-unique.js";
+export { getPrefix, isId } from "./utils/validate.js";
 
-export type { IdGenerator, IdOptions, PrefixedId } from "./types.js";
-export type { EnsureUniqueOptions } from "./ensure-unique.js";
-export type { TemplateOptions } from "./template.js";
+export type { IdGenerator, IdOptions, PrefixedId } from "./types/index.js";
+export type {
+  GetTimestampOptions,
+  SortableIdOptions,
+} from "./generators/sortable.js";
+export type { EnsureUniqueOptions } from "./utils/ensure-unique.js";
+export type { TemplateOptions } from "./generators/template.js";